add numbers and print results and count upto 98 functions

+++ 0x02-functions_nested_loops/main.h
@@ -56,7 +56,7 @@ void times_table(void);
 
 int add(int, int);
 
-void print_digits(long n);
+void print_digits(int n);
 
 void print_to_98(int n);
